primerfit-api/requirements.txt
APScheduler==3.6.3
asgiref==3.3.0
Babel==2.8.0
block-disposable-email==2.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
Django==2.2.10
django-cors-headers==3.2.1
django-disposable-email-checker==2.0.2
django-extensions==1.7.1
django-filter==2.0.0
django-phonenumber-field==4.0.0
django-rest-knox==3.0.3
django-rest-logger @ git+https://github.com/Seedstars/django-rest-logger.git@12f9b82edcde2a17a05bff3ed654187252b7e4cb
django-rest-swagger==2.2.0
django-rq==2.3.2
django-simple-history==2.12.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
phonenumbers==8.11.2
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.6.1
pytz==2020.4
raven==6.1.0
redis==3.5.3
requests==2.24.0
rq==1.5.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sentry-sdk==0.14.4
simplejson==3.17.2
six==1.10.0
sqlparse==0.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.11
whitenoise==3.3.0
gunicorn==19.7.1
dj-database-url==0.5.0
django-rest-multiple-models==2.1.3
django-rest-passwordreset==1.1.0